requirements.txt
fastapi
uvicorn
jinja2
pandas
python-multipart
openpyxl
requests
python-dotenv
